Add lti registration conversion to Canvas' format
This commit implements the Lti::ToolConfiguration "interface" in the
Lti::IMS::Regstration class. This allows the DeveloperKey class
to return a Lti::IMS::Registration object, instead of a
Lti::ToolConfiguration when it is based on the former.

test plan:
Install a Dynamic Registration tool, enable the developer key, and make
sure it is installed in the correct places and is launchable.

fixes INTEROP-7954
flag=lti_dynamic_registration

Change-Id: Ib5602869412dadd06007f278ad3fe29e99e29c28
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/331496
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Tucker Mcknight <tmcknight@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>