custom status total grade student summary page
this commit fixes the custom grade status not being displayed in the
student grade summary page when there is not an override score along
with the custom status. it also displays the total grade as '-' if the
allow_final_grade_value boolean value is true on the custom status. this
is only a presentation/display change and underlying grades are still
stored on the backend.

closes EVAL-3793
flag=custom_gradebook_statuses

test plan:
- Create a new custom status for a course.
- In a rails console run:
  CustomGradeStatus
  .find(id)
  .update(applies_to_submissions: false, allow_final_grade_value: false)
- Assign the custom status to a student's override score.
- Go to the student's grade summary page.
- Verify that the custom status is displayed in the total grade column.
- Verify that the total is '-'.

Change-Id: Iff44bac89783f56acc6c61e08b8742cd54e105eb
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/335300
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Rohan Chugh <rohan.chugh@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Rohan Chugh <rohan.chugh@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>