Don't assume the tool tray is for a module_index_menu launch
Previously we only checked if the placement was
equal to "module_index_menu" and passed that boolean
to the setExternalToolTray function.

But we actually do care about the exact placement we are
using, since this component is used in places other than
the module_index_menu. So let's pass in the placement
to the setExternalToolTray function, and then just check
if the placement is "module_index_menu" inside of the function.

Default to module_index_menu for all of the other uses of
setExternalToolTray that don't explicitly pass in their placement.

fixes INTEROP-8373

test plan:
- Install a test tool with the module_group_menu placement
- Go to the modules page and create a module
- Click the three-dots menu on your module and select
  your tool. (If there are multiple ones, choose the
  one that says "group menu.")
- When the tool shows up in the sidebar, check that the
  placement parameter given to it was "module_group_menu."
- Check that other placements on the modules page (like the
  non-group-menu ones) still work.
- Also make sure that there aren't any other uses of
  setExternalToolTray that have the third argument passed in.
  I looked, but it would be nice to double-check.

Change-Id: I0ad95475059ae643789e378412a091645ffbc3cd
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/337628
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>