remove custom fetch:fail event from our Backbone.Collection
the PaginatedCollectionView component that the event was
meant for (to disable its tiny loading spinner on xhr fail)
does not even listen for the fetch:fail event, leaving that
tiny loading spinner visible after it displays
a flash error message when there is an xhr failure; this was
the state prior to this changset since at least April 2023.

Since our custom canvas patch of Backbone.Collection to
fire this event is blocking the jquery 1.7 -> 1.8 upgrade,
the loading spinnner is tiny, the flash error msg is shown
regardless, and it hasn't been working already for quite
a while, this custom event is just being removed entirely.

If someone needs it back again it can be added in a way
that doesn't mess with jquery's quasi $.Deferred promises.

refs FOO-4220

flag = none

Test Plan: tests pass

Change-Id: I1069dc95abe52ad0aeaa7c2e16c014315e99def8
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/338068
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Jen Smith <jen.smith@instructure.com>
Product-Review: Jen Smith <jen.smith@instructure.com>