Fix order of rendering js_env for InstUI top_nav
Refs FOO-4126
Refs FOO-4127
flag=instui_nav

The earlier work to provide the new InstUI-based top nav
bar works fine, but when trying to add the new top nav
to another page, I ran into a snag. If the page being built
is rendered explicitly by the controller, and not by the
default Rails controller -> view logic flow, then things
happen out of order and the ENV.breadcrumbs never gets set
on the front end.

I'm not exactly sure of the flow reasoning of why it works
sometimes and not others, but the basic problem is that the
code to set ENV is emitted by the _head partial layout,
which runs prior to the main ERB view for the page itself.
And since that's where we were conditionally adding the
breadcrumbs to the env, it gets added too late to appear
in the ENV.

So this commit puts all the actions that need to happen when
we're rendering the new InstUI top nav into the controller,
and the view simply does the conditional rendering of the
old vs new HTML.

Test plan:
* repeat test plan for g/338376

Change-Id: Id3b807c8bbaf5353773bcb3b1233e06309872e20
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/340531
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Reviewed-by: Martin Ochoa <martin.ochoa@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>