Support no placements in Dynamic Registration
why

When a tool does not specify any placements, Canvas should default to
the `link_selection` placement. Previously, Canvas would render an error

This commit defaults the placement to `link_selection` when no placement
is specified in an LTI Message configuration. It also updates the
registration controller to remove duplicate placements. Also included in
this commit is a refactoring of the frontend to be based on the derived
native canvas configuration format. This is because the BE formats the
registration into the native canvas format, and the frontend was also

This means that if two LTI messages are defined like:

```
...
messages: [
  {type: "LtiResourceLinkRequest", url: "http://tool.com/rl"},
  {type: "LtiDeepLinkingRequest", url: "http://tool.com/dl"},
]
```

They will both default to `link_selection`, and only the first will be
included, due to de-duplication.

Also included in this commit is a refactoring of the frontend to be
based on the derived native canvas configuration format. This is because
the BE formats the registration into the native canvas format, and the
frontend was also duplicating that logic.

test plan:

Modify yaltt to not include any placements in an LTI Message. To do
this, you can comment out this line:

https://github.com/pfgray/yaltt/blob/6d70970a47bac6386f9991fe18f4aba0d62423b6/packages/frontend/src/views/lists/registrations/DynamicRegistration.tsx#L243

Then, install yaltt via Dynamic Registration. You should see the highest
placement chosen, and the placement should be `link_selection`.

Validate that the form works (including the "Reset to Defaults" button),
and that the registration is saved correctly.

fixes INTEROP-8425
flag=lti_dynamic_registration

Change-Id: Ie58f8e0a66fadfc752662a7632cdbc8f5a934eaf
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/340751
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Product-Review: Mark Starkman <mark.starkman@instructure.com>