Remove deprecated delegate() method from jquery datepickers
Refs FOO-4224
flag=none

The .delegate() way of installing event handlers has been
deprecated since jQuery 1.7 and is removed entirely in 3.x.

Canvas uses this in a bunch of places but we can eat away at
it a little at a time. This replaces it in the jquery and
jqueryUI datepickers.

Test plan:
* automated tests pass
* make sure uses of the old jQuery datepicker still work

Change-Id: Ib12d62cac5e94e511a9bc3caa5da1a5ffaa7d049
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/341012
Reviewed-by: Jen Smith <jen.smith@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>