update spec for react 18
This PS makes many of the failures in this file go away.

There is an issue with the kebob on the file page that will be
written up in a separate JIRA ticket.

Closes LF-1519

Test Plan: passes Jenkins

Change-Id: I694fcbf80757a243fa7f3491acda69e8e8db851b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/346557
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
Reviewed-by: Eric Saupe <eric.saupe@instructure.com>
QA-Review: Eric Saupe <eric.saupe@instructure.com>
Product-Review: Robin Kuss <rkuss@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>