Reassign assignment button available
Reassign assignment button is not available
immediately when an attachment is added
to a submission comment because author id
is an integer and is compared with an string

fixes EVAL-3987
flag=none

test plan:
- Create and publish an assignment in the course
with at least 2 attempts and a due date in the future.
- As a student make a submission.
- As the teacher navigate to SpeedGrader.
- Add a comment and add an attachment as well.
- Observe that the Reassign Assignment button
is now able.

Change-Id: I060ef18bb5d0a2cac9d4b97ac1609baa1eaa3de5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/347183
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>
Reviewed-by: Derek Williams <derek.williams@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Derek Williams <derek.williams@instructure.com>