Manage Apps Page: hooking up FE to BE
This commit hooks up the Manage Apps page to the backend. It fetches the
list of apps and displays them in a table. It also adds pagination and
search query parameters to the request, which will start working when
the backend supports it.

test plan:

Load the Apps page and see the list of apps.

closes INTEROP-8578
flag=lti_registrations_page

Change-Id: Ida676cf9813a3b64c0627531984eea2aeb77ebeb
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/350048
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>