Ensure EDD works with multiple assignments/module configurations
We previously fixed some bugs with EDD being run with multiple
assignments and module overrides being applied to every assignment
instead of just the correct assignments. This was a similar bug,
except if one assignment is in a module and another is not, the
SQL that checks if an assignment is in a module incorrectly calculated
all assignments as being in a module.

flag=selective_release_backend
closes LX-1729

test plan:
- have a course with assignments and modules
- test out different combinations of assignments
  being in modules, with or without overrides and run EDD
- EffectiveDueDates.for_course(course, assignment1, assignment2).to_hash
- Have one assignment not be in a module, and one be in a module
- have one be in a module with overrides, the other in an module with
  no overrides
- have both be in modules, both not be in modules, etc
- any other scenarios you can think of
- ensure that the EDD hash matches who is assigned to the assignment

Change-Id: I35a583516256551f790e246cf5dca120cdca04be
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/350101
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Sarah Gerard <sarah.gerard@instructure.com>