Abort help_links fetch when modal is canceled first
Refs FOO-4518
flag=none

This probably isn't strictly necessary, but I wanted
to explore the ability for an API call made by React
Query to be aborted in flight if the owning component
is unmounted. This turns out to be much easier than I
was afraid it would be; Tanstack Query has cancelable
query functionality built right in, and all that has
to be done is to pass along the relevant AbortSignal
option to doFetchApi.

I've already instrumented this all up with various
artificial delays, event handlers, and console logging
to verify this functionality, and it works great. Hit
me up for details if you want.

Also removed the recursion from the paginated loader,
seems unnecessary to do it that way when iteration
is so much more natural here.

Test plan:
* automated tests pass

Change-Id: I4100776709baf93dc75e62acd37b7503d622454a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/351520
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>