Streamline help_links query function
Refs FOO-4518
flag=none

There's no reason to manually return a Promise and then
explicitly handle its resolve/reject logic.  Tanstack
Query's docmentation makes it clear that the rejected
Promise caused by an async function throwing an error is
exactly what it is looking for as a signal that the query
failed. And since it is already the case that doFetchApi
will throw if the fetch response is not "ok" we don't need
to do anything beyond that. This greatly simplifies the
query function!

Test plan:
* automated tests pass
* help dialog looks right on both the login screen
  and on the Help Tray after login

Change-Id: I43f83fcc9fe03361351a6b44fe7dfd2d11a24693
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/351594
Reviewed-by: Michael Hulse <michael.hulse@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>