Fix 500 Error when sending Messages
closes OUT-6530
flag=none

We are getting 500 errors because MSW does not initiate
conversations with participants when sending messages. This
is because they are bulk messages. We do not want to trigger
OOO messages for bulk conversations because that may not be
a good experience. The solution is to mark the root_message
as 'automated' so there is no checking for OOO responses

Test Plan:
- Jenkins Tests Pass
- Go to any course with students
- Navigate to Traditional Gradebook
- Open 'Message Students Who...' for an assignment column
- Ensure that the message has at least one recipient
- Write a message and send
- Verify that there is no 500 error when sending
- Repeat for Assignment Group Column and Total Grade column

Change-Id: I64370fdb8f9ad13759d6bcf91df3fc1263e1dd3a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/352575
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Martin Yosifov <martin.yosifov@instructure.com>
QA-Review: Martin Yosifov <martin.yosifov@instructure.com>
Product-Review: Kyle Rosenbaum <krosenbaum@instructure.com>