fix eportfolio exports with long filenames
We are still seeing file names that exceed 255 characters when
it came time to processing the zip file. This was due to the fact that
the filenames get appended to for uniqueness etc., but we
were validating too far up the stack.

The approach also keeps the original section / page names up to
a point, but then truncates them if exceeding the limit, while
maintaining the extension names.

fixes FOO-4612
flag = none

test plan |
 - create a _long_ ePortfolio section name
 - create a _long_ ePortfolio page name within that section
 - both should exceed 255 characters when combined
 - click "Download the contents of this ePortfolio as a zip file"
 - the zip file should be created without error

Change-Id: Ib855a21ca1e428b7b3b3e81691200ec11b40d67d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/354739
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Michael Hulse <michael.hulse@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>